# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface LoginResponse {
  token: string;
}

export interface VerifyResponse {
  valid: boolean;
}

export type LifecycleUpdateState = typeof LifecycleUpdateState[keyof typeof LifecycleUpdateState];


export const LifecycleUpdateState = {
  draft: 'draft',
  published: 'published',
  archived: 'archived',
} as const;

export interface LifecycleUpdate {
  state: LifecycleUpdateState;
}

export type ApiProductPublishedState = typeof ApiProductPublishedState[keyof typeof ApiProductPublishedState];


export const ApiProductPublishedState = {
  draft: 'draft',
  published: 'published',
  archived: 'archived',
} as const;

export interface ApiProduct {
  id: number;
  name: string;
  category: string;
  tagline: string;
  description: string;
  version: string;
  status: string;
  features: string[];
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  apkUrl?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  /** @nullable */
  telegramUrl?: string | null;
  /** @nullable */
  screenshot1Url?: string | null;
  /** @nullable */
  screenshot2Url?: string | null;
  /** @nullable */
  screenshot3Url?: string | null;
  /** @nullable */
  screenshot4Url?: string | null;
  accentColor: string;
  publishedState: ApiProductPublishedState;
  sortOrder: number;
}

export type ProductInputPublishedState = typeof ProductInputPublishedState[keyof typeof ProductInputPublishedState];


export const ProductInputPublishedState = {
  draft: 'draft',
  published: 'published',
  archived: 'archived',
} as const;

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  tagline: string;
  /** @minLength 1 */
  description: string;
  /** @minLength 1 */
  version: string;
  status: string;
  features: string[];
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  apkUrl?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  /** @nullable */
  telegramUrl?: string | null;
  /** @nullable */
  screenshot1Url?: string | null;
  /** @nullable */
  screenshot2Url?: string | null;
  /** @nullable */
  screenshot3Url?: string | null;
  /** @nullable */
  screenshot4Url?: string | null;
  accentColor: string;
  publishedState?: ProductInputPublishedState;
  sortOrder: number;
}

export type ProductUpdatePublishedState = typeof ProductUpdatePublishedState[keyof typeof ProductUpdatePublishedState];


export const ProductUpdatePublishedState = {
  draft: 'draft',
  published: 'published',
  archived: 'archived',
} as const;

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  category?: string;
  /** @minLength 1 */
  tagline?: string;
  /** @minLength 1 */
  description?: string;
  /** @minLength 1 */
  version?: string;
  status?: string;
  features?: string[];
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  apkUrl?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  /** @nullable */
  telegramUrl?: string | null;
  /** @nullable */
  screenshot1Url?: string | null;
  /** @nullable */
  screenshot2Url?: string | null;
  /** @nullable */
  screenshot3Url?: string | null;
  /** @nullable */
  screenshot4Url?: string | null;
  accentColor?: string;
  publishedState?: ProductUpdatePublishedState;
  sortOrder?: number;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export interface ErrorResponse {
  error: string;
}

export interface DeletedResponse {
  success: boolean;
}