# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Admin authentication
  - name: products
    description: Product management
  - name: storage
    description: Object storage

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    LoginResponse:
      type: object
      properties:
        token:
          type: string
      required:
        - token

    VerifyResponse:
      type: object
      properties:
        valid:
          type: boolean
      required:
        - valid

    LifecycleUpdate:
      type: object
      properties:
        state:
          type: string
          enum: [draft, published, archived]
      required:
        - state

    ApiProduct:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        tagline:
          type: string
        description:
          type: string
        version:
          type: string
        status:
          type: string
        features:
          type: array
          items:
            type: string
        logoUrl:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        apkUrl:
          type: ["string", "null"]
        githubUrl:
          type: ["string", "null"]
        telegramUrl:
          type: ["string", "null"]
        screenshot1Url:
          type: ["string", "null"]
        screenshot2Url:
          type: ["string", "null"]
        screenshot3Url:
          type: ["string", "null"]
        screenshot4Url:
          type: ["string", "null"]
        accentColor:
          type: string
        publishedState:
          type: string
          enum: [draft, published, archived]
        sortOrder:
          type: integer
      required:
        - id
        - name
        - category
        - tagline
        - description
        - version
        - status
        - features
        - accentColor
        - publishedState
        - sortOrder

    ProductInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
          minLength: 1
        tagline:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        version:
          type: string
          minLength: 1
        status:
          type: string
        features:
          type: array
          items:
            type: string
        logoUrl:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        apkUrl:
          type: ["string", "null"]
        githubUrl:
          type: ["string", "null"]
        telegramUrl:
          type: ["string", "null"]
        screenshot1Url:
          type: ["string", "null"]
        screenshot2Url:
          type: ["string", "null"]
        screenshot3Url:
          type: ["string", "null"]
        screenshot4Url:
          type: ["string", "null"]
        accentColor:
          type: string
        publishedState:
          type: string
          enum: [draft, published, archived]
        sortOrder:
          type: integer
      required:
        - name
        - category
        - tagline
        - description
        - version
        - status
        - features
        - accentColor
        - sortOrder

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
          minLength: 1
        tagline:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        version:
          type: string
          minLength: 1
        status:
          type: string
        features:
          type: array
          items:
            type: string
        logoUrl:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        apkUrl:
          type: ["string", "null"]
        githubUrl:
          type: ["string", "null"]
        telegramUrl:
          type: ["string", "null"]
        screenshot1Url:
          type: ["string", "null"]
        screenshot2Url:
          type: ["string", "null"]
        screenshot3Url:
          type: ["string", "null"]
        screenshot4Url:
          type: ["string", "null"]
        accentColor:
          type: string
        publishedState:
          type: string
          enum: [draft, published, archived]
        sortOrder:
          type: integer

    RequestUploadUrlBody:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
      required:
        - name
        - size
        - contentType

    RequestUploadUrlResponse:
      type: object
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
      required:
        - uploadURL
        - objectPath

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    DeletedResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: JWT token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Missing fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/verify:
    get:
      operationId: verifyAuth
      tags: [auth]
      summary: Verify JWT token validity
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Token is valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyResponse"
        "401":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List published products (public)
      responses:
        "200":
          description: Published products only
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiProduct"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created (defaults to draft state)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiProduct"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/products:
    get:
      operationId: adminListProducts
      tags: [products]
      summary: List all products regardless of state (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All products (draft, published, archived)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiProduct"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiProduct"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiProduct"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /products/{id}/lifecycle:
    patch:
      operationId: updateProductLifecycle
      tags: [products]
      summary: Update product lifecycle state (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LifecycleUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiProduct"
        "400":
          description: Invalid state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadUrlBody"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestUploadUrlResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"