# lib/api-zod/src/generated/types/apiProduct.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { ApiProductPublishedState } from './apiProductPublishedState';

export interface ApiProduct {
  id: number;
  name: string;
  category: string;
  tagline: string;
  description: string;
  version: string;
  status: string;
  features: string[];
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  apkUrl?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  /** @nullable */
  telegramUrl?: string | null;
  /** @nullable */
  screenshot1Url?: string | null;
  /** @nullable */
  screenshot2Url?: string | null;
  /** @nullable */
  screenshot3Url?: string | null;
  /** @nullable */
  screenshot4Url?: string | null;
  accentColor: string;
  publishedState: ApiProductPublishedState;
  sortOrder: number;
}