# lib/api-zod/src/generated/types/lifecycleUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { LifecycleUpdateState } from './lifecycleUpdateState';

export interface LifecycleUpdate {
  state: LifecycleUpdateState;
}