# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { ProductUpdatePublishedState } from './productUpdatePublishedState';

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  category?: string;
  /** @minLength 1 */
  tagline?: string;
  /** @minLength 1 */
  description?: string;
  /** @minLength 1 */
  version?: string;
  status?: string;
  features?: string[];
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  apkUrl?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  /** @nullable */
  telegramUrl?: string | null;
  /** @nullable */
  screenshot1Url?: string | null;
  /** @nullable */
  screenshot2Url?: string | null;
  /** @nullable */
  screenshot3Url?: string | null;
  /** @nullable */
  screenshot4Url?: string | null;
  accentColor?: string;
  publishedState?: ProductUpdatePublishedState;
  sortOrder?: number;
}