# lib/api-zod/src/generated/types/requestUploadUrlBody.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}